OO refactor source code  (#427)
* Updates to the SExtractor-like catalog generation code.

* [WIP] -- Refactor HAP code to be object-oriented.

* Delete obsolete files and update pollers

* Adding modules with lower-case names

* [WIP] Updates to object-oriented implementation of the HAP portion of the package

* Small cleanups and removed runhlaprocessing.py and pipeline_poller_utils.py from repo
